requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
cython
cookiecutter
numpy
scipy
pandas
bloscpack
scikit-learn
scikit-image
xgboost
matplotlib
seaborn
pyyaml
wheel
pygments
ipython
jupyter
